Implemented basic drink state & prefab changes
The basic drink state searches for the closest water position from a list of empty objects. It them moves the animal to that object and increases thirst.

I also created all the prefabs I could think of that this project might need:
- Wolf
- Female wolf
- Baby wolf
- Baby deer

These and the previous prefabs are now seen as handlers as they contain a parent object with the nav mesh agent and relevant scripts.